+++ src/common/concurrent/thread_pool.rs
@@ -5,7 +5,7 @@ use std::{collections::HashMap, sync::Arc};
 
 static REGISTRY: Lazy<ThreadPoolRegistry> = Lazy::new(ThreadPoolRegistry::default);
 
-#[derive(Clone, Copy, Hash, PartialEq, Eq)]
+#[derive(Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
 #[cfg_attr(any(feature = "sync", feature = "future"), derive(Debug))]
 pub(crate) enum PoolName {
     Housekeeper,
@@ -31,6 +31,17 @@ pub(crate) struct ThreadPool {
     pub(crate) name: PoolName,
     pub(crate) pool: ScheduledThreadPool,
     // pub(crate) num_threads: usize,
+}
+
+impl ThreadPool {
+    fn new(name: PoolName, num_threads: usize) -> Self {
+        let pool = ScheduledThreadPool::with_name(name.thread_name_template(), num_threads);
+        Self {
+            name,
+            pool,
+            // num_threads,
+        }
+    }
 }
 
 pub(crate) struct ThreadPoolRegistry {
@@ -64,14 +75,8 @@ impl ThreadPoolRegistry {
                     // https://github.com/moka-rs/moka/pull/39#issuecomment-916888859
                     // https://github.com/seanmonstar/num_cpus/issues/69
                     let num_threads = num_cpus::get().max(1);
-                    let pool =
-                        ScheduledThreadPool::with_name(name.thread_name_template(), num_threads);
-                    let t_pool = ThreadPool {
-                        name,
-                        pool,
-                        // num_threads,
-                    };
-                    Arc::new(t_pool)
+                    let pool = ThreadPool::new(name, num_threads);
+                    Arc::new(pool)
                 });
             }
         }
@@ -89,5 +94,12 @@ impl ThreadPoolRegistry {
                 }
             }
         }
+    }
+
+    #[cfg(all(test, feature = "sync"))]
+    pub(crate) fn enabled_pools() -> Vec<PoolName> {
+        let mut names: Vec<_> = REGISTRY.pools.read().keys().cloned().collect();
+        names.sort_unstable();
+        names
     }
 }