+++ src/sync/builder.rs
@@ -56,6 +56,7 @@ pub struct CacheBuilder<K, V, C> {
     time_to_live: Option<Duration>,
     time_to_idle: Option<Duration>,
     invalidator_enabled: bool,
+    thread_pool_enabled: bool,
     cache_type: PhantomData<C>,
 }
 
@@ -76,6 +77,8 @@ where
             time_to_live: None,
             time_to_idle: None,
             invalidator_enabled: false,
+            // TODO: Change this to `false` in Moka 0.10.0.
+            thread_pool_enabled: true,
             cache_type: Default::default(),
         }
     }
@@ -117,6 +120,7 @@ where
             time_to_live: self.time_to_live,
             time_to_idle: self.time_to_idle,
             invalidator_enabled: self.invalidator_enabled,
+            thread_pool_enabled: self.thread_pool_enabled,
             cache_type: PhantomData::default(),
         }
     }
@@ -145,6 +149,7 @@ where
             self.time_to_live,
             self.time_to_idle,
             self.invalidator_enabled,
+            builder_utils::housekeeper_conf(self.thread_pool_enabled),
         )
     }
 
@@ -174,6 +179,7 @@ where
             self.time_to_live,
             self.time_to_idle,
             self.invalidator_enabled,
+            builder_utils::housekeeper_conf(self.thread_pool_enabled),
         )
     }
 }
@@ -208,6 +214,7 @@ where
             self.time_to_live,
             self.time_to_idle,
             self.invalidator_enabled,
+            builder_utils::housekeeper_conf(self.thread_pool_enabled),
         )
     }
 
@@ -238,6 +245,7 @@ where
             self.time_to_live,
             self.time_to_idle,
             self.invalidator_enabled,
+            builder_utils::housekeeper_conf(true),
         )
     }
 }
@@ -377,6 +385,22 @@ impl<K, V, C> CacheBuilder<K, V, C> {
     pub fn support_invalidation_closures(self) -> Self {
         Self {
             invalidator_enabled: true,
+            ..self
+        }
+    }
+
+    /// Specify whether or not to enable the thread pool for housekeeping tasks.
+    /// These tasks include removing expired entries and updating the LRU queue and
+    /// LFU filter. `true` to enable and `false` to disable. (Default: `true`)
+    ///
+    /// If disabled, the housekeeping tasks will be executed by a client thread when
+    /// necessary.
+    ///
+    /// NOTE: The default value will be changed to `false` in a future release
+    /// (v0.10.0 or v0.11.0).
+    pub fn thread_pool_enabled(self, v: bool) -> Self {
+        Self {
+            thread_pool_enabled: v,
             ..self
         }
     }